--- src/components/CommandInterface/CommandInterfaceDivider.tsx
@@ -1,3 +1,0 @@
-const CommandIntefaceDivider = () => <div className="h-0.5 bg-gray-600" />
-
-export default CommandIntefaceDivider
