+++ src/components/pure/ProjectsTable/index.tsx
@@ -5,8 +5,6 @@ import Loading from '@/components/pure/Loading'
 import TopBar from '@/components/page/overview/TopBar'
 import columns from '@/components/pure/ProjectsTable/columns'
 import FilterBar from '@/components/page/overview/Filterbar'
-import { useEffect, useState } from 'react'
-import CommandInterface from '@/components/commandinterface/CommandInterface'
 import Table from '@/components/page/overview/Table'
 
 const nullFunc = () => null
@@ -18,7 +16,6 @@ const ProjectsTable = () => {
   // Fetch data from Supabase using generated Urql hook
   const [{ data, fetching, error }] = useTrendingProjectsQuery()
   const projects = data?.projectCollection?.edges?.map((edge) => edge.node) as Project[]
-  const [openModal, setOpenModal] = useState<boolean>(false)
 
   // Initialize TanStack table
   const table = useReactTable({
@@ -26,26 +23,6 @@ const ProjectsTable = () => {
     columns,
     getCoreRowModel: getCoreRowModel()
   })
-
-  useEffect(() => {
-    function handleKeyDown(event: KeyboardEvent) {
-      if ((event.ctrlKey || event.metaKey) && event.key === 'k') {
-        event.preventDefault()
-        setOpenModal(true)
-      } else if (event.key === 'Escape') {
-        setOpenModal(false)
-      }
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown)
-    }
-  }, [])
-  const closeModel = () => {
-    setOpenModal(false)
-  }
   // Display loading/ error messages conditionally
   if (fetching) return <Loading message="Getting trending projects for you..." />
   if (!projects || projects.length === 0 || error) return <Error />
@@ -54,7 +31,6 @@ const ProjectsTable = () => {
     <div className="flex w-full flex-col">
       <TopBar columns={table.getAllLeafColumns()} nullFunc={nullFunc} />
       <FilterBar />
-      {openModal && <CommandInterface action={closeModel} />}
       <Table table={table} />
     </div>
   )
