+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import Page from '@/components/side-effects/Page'
 import withAuth from '@/components/side-effects/withAuth'
-
 import ProjectsTable from '@/components/pure/ProjectsTable'
 
 const Home = () => (
